+++ car_rental/src/main/java/services/ClientService.java
@@ -4,8 +4,8 @@ import entities.Client;
 
 public interface ClientService {
     void addClient(Client client);
-    void deleteClient(int id);
-    Client getClient(int id);
+    void deleteClient(long id);
+    Client getClient(long id);
     void updateClient(Client client);
     void changePassword(String oldPassword, String newPassword);
 }
