+++ car_rental/src/main/java/services/RepairInfoService.java
@@ -10,8 +10,8 @@ import java.util.Date;
 public interface RepairInfoService {
     void updateDateStart(Date date, RepairInfo repairInfo);
     void updateStatus(RepairInfo repairInfo, RepairStatus repairStatus);
-    RepairInfo getRepairInfoById(int id);
-    Vehicule getVehiculeById(int id);
-    Repair getRepair(int id);
+    RepairInfo getRepairInfoById(long id);
+    Vehicule getVehiculeById(long id);
+    Repair getRepair(long id);
 
 }
