+++ car_rental/src/main/java/services/VehiculeService.java
@@ -8,7 +8,7 @@ public interface VehiculeService {
     void addVehicule(Vehicule vehicule);
     void deleteVehicule(Vehicule vehicule);
     void updateVehicule(Vehicule vehicule);
-    Vehicule getVehicule(int id);
-    Category getCategory(int id);
+    Vehicule getVehicule(long id);
+    Category getCategory(long id);
 
 }
